requirements.txt
# general modules
Django >4.2,<4.3
python-decouple
pillow
djangorestframework

# third party modules
markdown 
django-filter
drf-yasg[validation]
djangorestframework_simplejwt

# deployment modules
coreapi

# email third party modules
django-mail-templated

# refactoring and tests
black
flake8
pytest
pytest-django